# feat: enable gpu channel async from the runtime