@bpasero 2025-02-24
Just a guess, but maybe its because we specifically disable GPU for the perf machines?

https://github.com/microsoft/vscode-perf/blob/4a466010310f1dcd68206b6462f46eb8b6a43d17/src/perf.ts#L144-L150