@deepak1556 2025-02-26
Verified on my windows machine to show improvements with the exploration build similar to the numbers from https://github.com/microsoft/vscode/pull/241640#issue-2873429939

There is definitely something else blocking the startup path on the perf machines, I will capture traces from the perf bots as next step.